Finalizando os headers integrados
