+++ src/display/imagewidget.c
