+++ src/display/trackbar.c
