+++ src/imgproc/drawing.c
