--- src/type.c
@@ -1,37 +1,0 @@
-/* ===================================================================
-#   Copyright (C) 2015-2015
-#   Anderson Tavares <nocturne.pe at gmail.com> PK 0x38e7bfc5c2def8ff
-#   Lucy Mansilla    <lucyacm at gmail.com>
-#   Caio de Braz     <caiobraz at gmail.com>
-#   Hans Harley      <hansbecc at gmail.com>
-#   Paulo Miranda    <pavmbr at yahoo.com.br>
-#
-#   Institute of Mathematics and Statistics - IME
-#   University of Sao Paulo - USP
-#
-#   This file is part of Grafeo.
-#
-#   Grafeo is free software: you can redistribute it and/or
-#   modify it under the terms of the GNU General Public License
-#   as published by the Free Software Foundation, either version
-#   3 of the License, or (at your option) any later version.
-#
-#   Grafeo is distributed in the hope that it will be useful,
-#   but WITHOUT ANY WARRANTY; without even the implied warranty
-#   of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
-#   See the GNU General Public License for more details.
-#
-#   You should have received a copy of the GNU General Public
-#   License along with Grafeo.  If not, see
-#   <http://www.gnu.org/licenses/>.
-# ===================================================================*/
-#include <grafeo/type.h>
-
-int8_t int64_compare_function_r(void* value1, void* value2){
-  if((int64_t) value1 > (int64_t) value2) return -1;
-  else if((int64_t) value1 == (int64_t) value2) return 0;
-  else return 1;
-}
-int8_t int64_compare_function(void* value1, void* value2){
-  return -int64_compare_function_r(value1, value2);
-}
